// shelter-frontend/src/app/api/model/createFakturyDto.ts
/**
 * Shelter
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateFakturyDto {
    numer?: string;
    data?: string;
    kwota?: number;
    lecznicaId?: number;
    zwierzeta?: number[];
}